// apps/web/src/lib/generated/models/ClinicalTrialStatisticalAnalysisRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClinicalTrialStatisticalAnalysisRead = {
  confidence_interval_percent?: (number | null);
  lower_limit?: (number | null);
  method?: (string | null);
  notes?: (string | null);
  p_value?: (string | null);
  parameter_type?: (string | null);
  parameter_value?: (number | null);
  upper_limit?: (number | null);
};